requirements.txt
asn1crypto
pyyaml
requests
apkverify
tqdm